requirements.txt
pyaudio
grpcio
requests
soundfile
numpy
google-api-python-client